Improve device wake-up logic and debug messages.
The atsha204_i2c_wakeup() function did not work reliably. To wake the chip it
is necessary to hold SDA low for at least tWLO (60us), then delay for tWHI
(500us for ATECC108A; 2.5ms for ATSHA204). There is no way to explicitly hold
the SDA line low in Linux, so instead attempt to send 0 to address 0; this will
hold SDA low for 8 clock cycles (sending the address), which will be more than
60us if the I2C clock speed is less than 133kHz.

The limitation this imposes on the clock speed is restrictive, but this logic
does at least work reliably when the clock speed is under this threshold.

This commit also tidies up the debug so that dev_dbg() is used throughout, and
reorders some of the messages so that they come out in a logical order that
makes debugging easier.